Telebot/requirements.txt
python-telegram-bot==20
pandas==2.1.3